fix(RHICOMPL-1102): Deleting test result keeps policy profile
The test we had for this case relied on `external=true` rather than the
policy association existing, thus not catching the bug. I've updated the
test and the code to delete the profile only if no policy exists; for
internal policies, deleting the test results for any of the profiles
deletes test results for the entire policy's profiles.

Signed-off-by: Andrew Kofink <akofink@redhat.com>